Move Wordsy Keyboard and Grid out of utils
