+++ src/components/Game/Wordsy/Grid/EmptyRow.js
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 
 // FPCC
-import Cell from 'components/Game/Wordsy/Utils/Grid/Cell'
+import Cell from 'components/Game/Wordsy/Grid/Cell'
 
 function EmptyRow({ wordLength }) {
   const emptyCells = Array.from(Array(wordLength))
