+++ src/components/Game/Wordsy/WordsyPresentation.js
@@ -3,8 +3,8 @@ import PropTypes from 'prop-types'
 
 // FPCC
 import getIcon from 'common/utils/getIcon'
-import Grid from 'components/Game/Wordsy/Utils/Grid/Grid'
-import Keyboard from 'components/Game/Wordsy/Utils/Keyboard/Keyboard'
+import Grid from 'components/Game/Wordsy/Grid/Grid'
+import Keyboard from 'components/Game/Wordsy/Keyboard/Keyboard'
 import SectionTitle from 'components/SectionTitle'
 import InfoModal from 'components/Game/Wordsy/Modals/InfoModal'
 import EndGameModal from 'components/Game/Wordsy/Modals/EndGameModal'
